Move eval package into blockprocessor.
